Fix up caching imports
- Move the cfunc cache tests to test_caching
- Fix imports of BaseCacheTest in other tests
