+++ numba/tests/npyufunc/test_caching.py
@@ -6,7 +6,7 @@ import subprocess
 import numpy as np
 
 from numba.tests.support import capture_cache_log
-from numba.tests.test_dispatcher import BaseCacheTest
+from numba.tests.test_caching import BaseCacheTest
 from numba.core import config
 import unittest
